third_party/blink/web_tests/platform/generic/fast/forms/select/selected-false-expected.txt

1
2
3
PASS selectElement.selectedIndex = 0; selectElement.selectedIndex is 0
PASS optionElements[0].selected is true
PASS optionElements[1].selected is false
PASS optionElements[2].selected is false
PASS selectElement.selectedIndex = -1; selectElement.selectedIndex is -1
PASS optionElements[0].selected is false
PASS optionElements[1].selected is false
PASS optionElements[2].selected is false
PASS selectElement.selectedIndex = -2; selectElement.selectedIndex is -1
PASS selectElement.selectedIndex is 1
PASS optionElements[0].selected is false
PASS optionElements[1].selected is true
PASS optionElements[2].selected is false
PASS selectElement.selectedIndex is 2
PASS optionElements[0].selected is false
PASS optionElements[1].selected is false
PASS optionElements[2].selected is true
PASS selectElement.selectedIndex is 0
PASS optionElements[0].selected is true
PASS optionElements[1].selected is false
PASS optionElements[2].selected is false
PASS selectElement.selectedIndex is 1
PASS optionElements[0].selected is false
PASS optionElements[1].selected is true
PASS optionElements[2].selected is false
PASS successfullyParsed is true

TEST COMPLETE
